requirements.txt
Pillow==8.2.0
pandas==1.1.3
